% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-data.R
\name{create.data}
\alias{create.data}
\title{Generates random effects data}
\usage{
create.data(clusters, cluster.size, params.fe, params.re, G, ...)
}
\arguments{
\item{clusters}{The number of clusters (> 1)}

\item{cluster.size}{Either a vector indicating the size of each cluster or a single number which indicates the size of each cluster (same across clusters)}

\item{params.fe}{A list containing the values of the fixed effects}

\item{params.re}{A list contiaining the values of the random effects}

\item{G}{"True" mixing distribution as a character (e.g. "normal", "gamma")}

\item{...}{Named parameters for the mixing distribution G}
}
\value{
data.frame
}
\description{
Function can generate data with random intercept and fixed slope,
or random intercept and slope with arbitrary fixed covariates
}

